+++ ydb/core/persqueue/ut/internals_ut.cpp
@@ -1,5 +1,5 @@
 #include <ydb/core/persqueue/pqtablet/blob/blob.h>
-#include "header.h"
+#include <ydb/core/persqueue/pqtablet/blob/header.h>
 
 #include <library/cpp/testing/unittest/registar.h>
 #include <ydb/core/persqueue/partition_key_range/partition_key_range.h>
